add links to two new labs
Thank you all for creating such a wonderful resource! The LinuxONE Community Cloud was extremely helpful to me because I was leading two courses including hands-on labs. These labs are the hands-on portion of those courses, which stand on their own without the rest of the course.

I'm contributing them back here for other people to use, and because I noticed there weren't any labs in this folder yet.

If you would like me to contribute the entire project to this one instead of just adding the links to the README in this folder, let me know and I'm happy to do that as well.

Signed-off-by: Jacob Emery <jacob.emery@ibm.com>